Implement PayPalButtons by extracting a view.
